Moved server port into config and renamed index to server  npm start works correctly and closes #33
