+++ docs/index.md
@@ -6,5 +6,5 @@ See [configuration](./docs/configuration.md)
 ### Running the report
 
 ````
-GH_TOKEN=<token> node index.js
+npm start
 ````
